worker/requirements.txt
datalab-python-sdk>=0.1.0
flask>=3.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
